# web/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CUSA SSO API",
    "version": "1.4.0",
    "description": "Custom first-party SSO: Authorization Code + PKCE S256, opaque tokens, application-scoped roles. Not a full OAuth/OIDC provider; no discovery, ID token, refresh token or dynamic registration. Examples are placeholders; never paste production secrets into browser documentation."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin as this documentation; production uses HTTPS"
    }
  ],
...
  "paths": {
...
    "/api/admin/applications/{id}/sharing": {
      "get": {
        "operationId": "getSharingPolicy",
        "summary": "ดูขอบเขตข้อมูลสูงสุดของ Service",
        "description": "CUSA first-party UI endpoint; full session required. Mutations require exact Origin and CSRF header. Never call from the service browser. Admin only. PUT requires fresh Passkey/TOTP and increments policy version, invalidating old requests/grants/tokens (cache bound applies). Scopes must include identity:read. Extra sensitive scopes are disabled by default.",
        "tags": [
          "Consent"
        ],
        "security": [
          {
            "CookieSession": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "scopes": {
                      "type": "array",
                      "maxItems": 7,
                      "uniqueItems": true,
                      "items": {
                        "type": "string",
                        "enum": [
                          "identity:read",
                          "profile",
                          "email",
                          "phone",
                          "phone:match",
                          "line",
                          "assurance"
                        ]
                      }
                    },
                    "purpose": {
                      "type": "string",
                      "description": "Concrete purpose shown to the user",
                      "minLength": 10,
                      "maxLength": 500
                    },
                    "version": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "scopes",
                    "purpose",
                    "version"
                  ],
                  "additionalProperties": false
                },
                "example": {
                  "scopes": [
                    "identity:read",
                    "profile",
                    "email"
                  ],
                  "purpose": "เข้าสู่ระบบและแสดงข้อมูลบัญชีใน Service นี้",
                  "version": 1
                }
              }
            }
          },
          "400": {
            "description": "Validation / invalid grant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Invalid request parameters",
                  "code": "invalid_request"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Invalid or expired API key",
                  "code": "invalid_client"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient scope / CORS / account access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "API key does not permit this operation",
                  "code": "insufficient_scope"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found in the permitted scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Resource not found",
                  "code": "NOT_FOUND"
                }
              }
            }
          },
          "409": {
            "description": "Policy version or activity identifier conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Reload the current policy or use the correct event identifier",
                  "code": "CONFLICT"
                }
              }
            }
          },
          "429": {
            "description": "Quota exceeded; respect Retry-After",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Service request quota exceeded",
                  "code": "RATE_LIMITED"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "เวลารอเป็นวินาที ห้าม retry รัว",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            }
          },
          "503": {
            "description": "Dependency or audit unavailable; fail closed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Service unavailable",
                  "code": "AUDIT_UNAVAILABLE"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Unable to process request",
                  "code": "INTERNAL_ERROR"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "saveSharingPolicy",
        "summary": "ตั้งค่าขอบเขตข้อมูลและวัตถุประสงค์",
        "description": "CUSA first-party UI endpoint; full session required. Mutations require exact Origin and CSRF header. Never call from the service browser. Admin only. PUT requires fresh Passkey/TOTP and increments policy version, invalidating old requests/grants/tokens (cache bound applies). Scopes must include identity:read. Extra sensitive scopes are disabled by default.",
        "tags": [
          "Consent"
        ],
        "security": [
          {
            "CookieSession": [],
            "CsrfToken": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SharingPolicy"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "const": true
                    }
                  },
                  "required": [
                    "ok"
                  ],
                  "additionalProperties": false
                },
                "example": {
                  "ok": true
                }
              }
            }
          },
          "400": {
            "description": "Validation / invalid grant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Invalid request parameters",
                  "code": "invalid_request"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Invalid or expired API key",
                  "code": "invalid_client"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient scope / CORS / account access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "API key does not permit this operation",
                  "code": "insufficient_scope"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found in the permitted scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Resource not found",
                  "code": "NOT_FOUND"
                }
              }
            }
          },
          "409": {
            "description": "Policy version or activity identifier conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Reload the current policy or use the correct event identifier",
                  "code": "CONFLICT"
                }
              }
            }
          },
          "429": {
            "description": "Quota exceeded; respect Retry-After",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Service request quota exceeded",
                  "code": "RATE_LIMITED"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "เวลารอเป็นวินาที ห้าม retry รัว",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            }
          },
          "503": {
            "description": "Dependency or audit unavailable; fail closed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Service unavailable",
                  "code": "AUDIT_UNAVAILABLE"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Unable to process request",
                  "code": "INTERNAL_ERROR"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Application-bound backend secret. Scope identity:read, token:introspect, token:revoke or member:activity as required."
      },
      "CookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-cusa_session",
        "description": "HttpOnly Secure cookie in production; cusa_session on localhost development"
      },
      "CsrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      },
      "BearerToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "ข้อความที่ปลอดภัยสำหรับแสดงผล"
          },
          "code": {
            "type": "string",
            "description": "Machine-readable error code; case-sensitive"
          },
          "requestId": {
            "type": "string",
            "description": "Server-generated correlation UUID ถ้ามี",
            "format": "uuid"
          }
        }
      },
...
      "SharingPolicy": {
        "type": "object",
        "properties": {
          "scopes": {
            "type": "array",
            "maxItems": 7,
            "uniqueItems": true,
            "items": {
              "type": "string",
              "enum": [
                "identity:read",
                "profile",
                "email",
                "phone",
                "phone:match",
                "line",
                "assurance"
              ]
            }
          },
          "purpose": {
            "type": "string",
            "description": "Concrete purpose shown to the user",
            "minLength": 10,
            "maxLength": 500
          }
        },
        "required": [
          "scopes",
          "purpose"
        ],
        "additionalProperties": false
      },
...
    }
  }
}